PWX-31198: record event using new eventsv1 interface (#317)
client-go has a new eventsv1 interface for recording events. This patch
adds a new method to record events. The old method is kept for backwards-compatibility
but it creates a new-style event by mapping the original parameters to new ones.
The original method that creates old-style events has been preserved by renaming
it with suffix "legacy".

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>